Deduplicate terminal buffers, saving ~700 bytes of RAM
History, line editor and input buffer are now single static inline
copies instead of per-instance/per-TU duplicates (script_terminal
carried ~530 dead bytes, input_buffer existed once per TU). The script
runner saves and restores the shared input buffer around each line so
a half-typed interactive command survives script execution, and script
action arguments are copied out of the buffer before restoring.
SoundNote/PatternStep use narrow types, halving the pattern buffers.

+++ code/tools.hpp
@@ -98,11 +98,13 @@ extern  void  sound_scaled(usize pin, isize freq_Hz, usize duration_ms, usize vo
 extern  void  sound_stop(void);
 
 // Асинхронный звуковой паттерн (команда beep): частота 0 - пауза.
+// Узкие типы намеренно: буфер паттерна живёт в ОЗУ постоянно (8 байт на ноту
+// вместо 16), а 65 кГц / 65 с на ноту хватает с запасом.
 struct SoundNote {
-  isize frequency_Hz;
-  usize duration_ms;
-  usize gap_ms;
-  usize volume_percent;
+  u16 frequency_Hz;
+  u16 duration_ms;
+  u16 gap_ms;
+  u8  volume_percent;
 };
 static constexpr usize SOUND_PATTERN_MAX = 16;
 extern  bool  sound_pattern_start(const SoundNote* notes, usize count);
